% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifires-package.R
\docType{package}
\name{identifires-package}
\alias{identifires-package}
\alias{identifires}
\title{identifires - identifiers.org Client}
\description{
identifires - identifiers.org Client
}
\section{identifiers.org API Docs}{

See \url{https://identifiers.org/restws}
}

\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
