% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_collections.R
\name{idf_collections}
\alias{idf_collections}
\title{Collections}
\usage{
idf_collections(ids = NULL, names = NULL, providers = NULL, ...)
}
\arguments{
\item{ids}{(character) one or more identifiers. optional.}

\item{names}{(character) one or more names. optional}

\item{providers}{(character) a provider prefix. optional}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
data.frame or list, depending on what's passed to the function
}
\description{
retrieves a list of all available collections
}
\details{
If no parameters passed, you get all collections.

When more than 1 string passed to \code{ids}, \code{names}, or \code{providers},
we use async http requests
}
\examples{
\dontrun{
(x <- idf_collections())

idf_collections(ids = "MIR:00000008")
ids <- c("MIR:00000249", "MIR:00000307", "MIR:00000308", "MIR:00000312", 
  "MIR:00000313", "MIR:00000315")
idf_collections(ids = ids)
idf_collections(names = "pdb")
idf_collections(names = "pdb")
idf_collections(providers = "pdb")
idf_collections(providers = "pdb")
}
}
