% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_resources.R
\name{idf_resources}
\alias{idf_resources}
\title{Resources}
\usage{
idf_resources(ids = NULL, ...)
}
\arguments{
\item{ids}{(character) one or more identifiers. optional.
when more than 1 passed, we use async http requests}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
When \code{ids} passed, data.frame of the form :
\itemize{
\item id
\item accessurl
\item info
\item institution
\item location
\item official
\item resourceprefix
\item localid
\item resourceurl
}

When \code{ids} NOT passed, data.frame of the form :
\itemize{
\item urlPrefix
\item urlSuffix
\item urlResource
\item info
\item institution
\item location
\item example
\item official
\item obsolete
\item frameDeny
\item convertPrefix
\item resourcePrefix
\item _links
}
}
\description{
retrieves the list of all available resources
}
\details{
if no parameters passed, gets all resources
}
\examples{
\dontrun{
x <- idf_resources()
idf_resources(ids = "MIR:00100675")
idf_resources(ids = c("MIR:00100675", "MIR:00100005"))
}
}
