% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_validate.R
\name{idf_validate}
\alias{idf_validate}
\title{Validate identifiers}
\usage{
idf_validate(ids, ...)
}
\arguments{
\item{ids}{(character) one or more identifiers. required.
when more than 1 passed, we use async http requests}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
When \code{ids} passed, data.frame of the form :
\itemize{
\item prefix
\item identifier
\item url
}
}
\description{
validates the given prefix:identifier scheme
}
\details{
if no parameters passed, gets all resources
}
\examples{
\dontrun{
idf_validate(ids = "MIR:00100675")
idf_validate(ids = c("MIR:00100675", "MIR:00100005"))
}
}
